Fixed application logic.
