--- Code/Engine/Application/Application.cpp
@@ -1,8 +1,0 @@
-/**
-* Application.cpp
-* 28.01.2023
-* (c) Denis Romakhov
-*/
-
-#include "pch.h"
-#include "Application/ApplicationInterface.h"
