+++ Code/Engine/Application/ApplicationInterface.h
@@ -14,9 +14,7 @@
 #include "GUI/Module.h"
 #include "World/Module.h"
 #include "RenderSystem/Module.h"
-#include "Core/ScriptSystem.h"
 #include "Core/LCTypes.h"
-#include "Core/Module.h"
 
 #pragma warning(disable : 4251)
 
@@ -39,7 +37,7 @@ typedef std::function<void(LcMouseBtn, LcKeyState, float, float, class IApplicat
 
 /**
 * Application interface */
-class APPLICATION_API IApplication
+class IApplication
 {
 public:
 	//
@@ -70,11 +68,14 @@ public:
 	* Set script system */
 	virtual void SetScriptSystem(TScriptSystemPtr scripts) noexcept = 0;
 	/**
-	* Set GUI manager */
-	virtual void SetGuiManager(TGuiManagerPtr gui) noexcept = 0;
+	* Set audio system */
+	virtual void SetAudioSystem(TAudioSystemPtr audio) noexcept = 0;
 	/**
 	* Set physics world */
 	virtual void SetPhysicsWorld(TPhysicsWorldPtr physWorld) noexcept = 0;
+	/**
+	* Set GUI manager */
+	virtual void SetGuiManager(TGuiManagerPtr gui) noexcept = 0;
 	/**
 	* Set window size in pixels */
 	virtual void SetWindowSize(int width, int height) = 0;
@@ -121,16 +122,22 @@ public:
 	* Get World pointer */
 	virtual TWorldPtr GetWorldPtr() noexcept = 0;
 	/**
-	* Get Physics World pointer */
-	virtual class IPhysicsWorld* GetPhysicsWorld() noexcept = 0;
-	/**
-	* Get Physics World pointer */
-	virtual TPhysicsWorldPtr GetPhysicsWorldPtr() noexcept = 0;
-	/**
 	* Get Script system pointer */
 	virtual class IScriptSystem* GetScriptSystem() noexcept = 0;
 	/**
 	* Get Script system pointer */
 	virtual TScriptSystemPtr GetScriptSystemPtr() noexcept = 0;
+	/**
+	* Get Audio system pointer */
+	virtual class IAudioSystem* GetAudioSystem() noexcept = 0;
+	/**
+	* Get Audio system pointer */
+	virtual TAudioSystemPtr GetAudioSystemPtr() noexcept = 0;
+	/**
+	* Get Physics World pointer */
+	virtual class IPhysicsWorld* GetPhysicsWorld() noexcept = 0;
+	/**
+	* Get Physics World pointer */
+	virtual TPhysicsWorldPtr GetPhysicsWorldPtr() noexcept = 0;
 
 };
