+++ Code/Engine/Core/LCTypesEx.h
@@ -11,6 +11,7 @@
 #include <algorithm>
 
 #ifdef _WINDOWS
+# include <windows.h>
 # include <directxmath.h>
 # include <dcommon.h>
 #endif
@@ -128,85 +129,3 @@ CORE_API LcMatrix4 TransposeMatrix(const LcMatrix4& mat);
 /**
 * Identity matrix */
 CORE_API LcMatrix4 IdentityMatrix();
-
-
-/** Default lifetime strategy */
-template<class T>
-class LcLifetimeStrategy
-{
-public:
-	LcLifetimeStrategy() {}
-	//
-	virtual ~LcLifetimeStrategy() {}
-	//
-	virtual std::shared_ptr<T> Create() { return std::shared_ptr<T>(); }
-	//
-	virtual void Destroy(T& item) {}
-};
-
-/** Object creator */
-template<class I, class T, class S = LcLifetimeStrategy<I>>
-class LcCreator : public ICreator<I>
-{
-public:
-	typedef std::shared_ptr<S> TStrategyType;
-	typedef std::deque<std::shared_ptr<I>> TCreatorItemsList;
-
-
-public:
-	LcCreator() : lifetimeStrategy(std::make_shared<S>()) {}
-	//
-	~LcCreator()
-	{
-		for (auto& item : items)
-		{
-			lifetimeStrategy->Destroy(*item.get());
-		}
-
-		items.clear();
-	}
-	//
-	template<class C>
-	C* AddT()
-	{
-		items.push_back(lifetimeStrategy->Create());
-		return static_cast<C*>(items.back().get());
-	}
-	//
-	virtual void SetLifetimeStrategy(TStrategyType strategy)
-	{
-		if (strategy) lifetimeStrategy = strategy;
-	}
-
-
-public: // ICreator interface implementation
-	//
-	virtual I* Add() override
-	{
-		return AddT<I>();
-	}
-	//
-	virtual void Remove(I* inItem) override
-	{
-		if (inItem)
-		{
-			auto it = std::find_if(items.begin(), items.end(), [inItem](const std::shared_ptr<I>& item) {
-				return item.get() == inItem;
-			});
-
-			lifetimeStrategy->Destroy(*inItem);
-			items.erase(it);
-		}
-	}
-	//
-	virtual const TCreatorItemsList& GetList() const override { return items; }
-	//
-	virtual TCreatorItemsList& GetList() override { return items; }
-
-
-protected:
-	TCreatorItemsList items;
-	//
-	TStrategyType lifetimeStrategy;
-
-};
