+++ Code/Engine/Core/ScriptSystem.h
@@ -29,8 +29,6 @@ public:
 
 };
 
-typedef std::shared_ptr<IScriptSystem> TScriptSystemPtr;
-
 
 /**
 * Script module */
