+++ Projects/Windows/Application/Application.vcxproj
@@ -154,7 +154,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\Code\Engine\Application\AppConfig.cpp" />
-    <ClCompile Include="..\..\..\Code\Engine\Application\Application.cpp" />
     <ClCompile Include="dllmain.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
