+++ Projects/Windows/Core/Core.vcxproj.filters
@@ -54,6 +54,9 @@
     <ClInclude Include="..\..\..\Code\Engine\Core\Audio.h">
       <Filter>Header Files\Core</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\Code\Engine\Core\LCCreator.h">
+      <Filter>Header Files\Core</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dllmain.cpp">
